Update OpenglParser.py
You will need to make multiple changes in the GLSLParser in order to add token recognition for else if. Here's how to accomplish it:

1.Revise the Lexer to Identify an alternative if: Make that the ELSEIF token is recognised by the lexer. Usually, the lexer definition file takes care of this.

2.Alter the Parser to Manage the remainder if: To handle the else if condition, modify the parse_if method.

The necessary adjustments are listed below:

Step 1: Update the Lexer (Already done)
Make sure that else if is recognised by the OpenglLexer as a single token, ELSEIF.
Step 2: Adjust the Interpreter
Boost token recognition for ELSEIF:
To identify and manage ELSEIF, you must provide a condition in the parse_if method.